Add center dashed line
Fixes positioning of sprites to comply with bevy's origin being in
the center of the window. This is the default behavior and it seems
easiest to cooperate with bevy.

Adds a dashed line down the center of the window. Dashed line is just a
batch of bevy SpriteBundles. As best as I can tell there is no way to
render a dashed-line using the existing bevy APIs. Positioning of the
dashed line is adaptive based on the window but uses pixels rather than
relative distances.

There is a slight misalignment I was not able to solve appropriately.
The correct implementation involves:
1. Defining a top/bottom "gutter" (space where no dashed line is
   rendered)
2. Rendering a dashed line on top and one on bottom
3. Computing the remaining space between lowest point of the top dash
   and the highest point of the bottom dash; then figuring out how many
   dashes can fit in between with equal spacing.

This layout is quite similar to `flex` in the browser. Since Rust was
written for Servo in Mozilla it only figures that we find ourselves at
this point again. Déjà vu.